Cambios en css y modificacion ruta del proyecto
