+++ Frontend/src/main/frontend/themes/frontend/catastrophe/addCatastrophe.css
@@ -4,17 +4,6 @@
     height: 100%;
 }
 
-/* Estilos para el header */
-.header {
-    display: flex;
-    align-items: center;
-    padding: 16px 24px;
-    background-color: #fff;
-    border-bottom: 1px solid #e0e0e0;
-    margin-bottom: 24px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
-}
-
 .title {
     margin: 0;
     font-size: 24px;
@@ -54,20 +43,6 @@
     background-color: #fff;
     border-radius: 8px;
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.05);
-}
-
-/* Campos del formulario */
-vaadin-text-field,
-vaadin-text-area,
-vaadin-date-picker,
-vaadin-number-field,
-vaadin-combo-box {
-    width: 100%;
-    --lumo-font-size-m: 16px;
-}
-
-vaadin-text-area {
-    min-height: 150px;
 }
 
 /* Contenedor de botones */
