--- Frontend/src/main/java/com/example/application/model/GPSCoordinates.java
@@ -1,19 +1,0 @@
-package com.example.application.model;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class GPSCoordinates {
-    private int id;
-
-    private double latitude;
-
-    private double longitude;
-
-    public GPSCoordinates(double latitude, double longitude) {
-        this.latitude = latitude;
-        this.longitude = longitude;
-    }
-}
