--- Frontend/src/main/java/com/example/application/model/enums/EmergencyLevel.java
@@ -1,5 +1,0 @@
-package com.example.application.model.enums;
-
-public enum EmergencyLevel {
-    LOW, MEDIUM, HIGH, VERYHIGH
-}
