--- Frontend/src/main/java/com/example/application/model/enums/NeedType.java
@@ -1,5 +1,0 @@
-package com.example.application.model.enums;
-
-public enum NeedType {
-    MEDICAL, POLICE, FIREFIGHTERS, CLEANING, FEED, PSYCHOLOGICAL, BUILDING, CLOTHING, REFUGE, OTHER
-}
