--- Frontend/src/main/java/com/example/application/model/enums/Priority.java
@@ -1,5 +1,0 @@
-package com.example.application.model.enums;
-
-public enum Priority {
-    URGENT, MODERATE, LOW
-}
