--- Frontend/src/main/java/com/example/application/model/enums/Status.java
@@ -1,5 +1,0 @@
-package com.example.application.model.enums;
-
-public enum Status {
-    TO_DO, IN_PROGRESS, FINISHED
-}
