--- Frontend/src/main/java/com/example/application/model/enums/UrgencyLevel.java
@@ -1,5 +1,0 @@
-package com.example.application.model.enums;
-
-public enum UrgencyLevel {
-    URGENT, MODERATE, LOW
-}
