+++ Frontend/src/main/java/solidarityhub/frontend/model/GPSCoordinates.java
@@ -1,0 +1,19 @@
+package solidarityhub.frontend.model;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class GPSCoordinates {
+    private int id;
+
+    private double latitude;
+
+    private double longitude;
+
+    public GPSCoordinates(double latitude, double longitude) {
+        this.latitude = latitude;
+        this.longitude = longitude;
+    }
+}
