+++ Frontend/src/main/java/solidarityhub/frontend/model/enums/EmergencyLevel.java
@@ -1,0 +1,5 @@
+package solidarityhub.frontend.model.enums;
+
+public enum EmergencyLevel {
+    LOW, MEDIUM, HIGH, VERYHIGH
+}
