+++ Frontend/src/main/java/solidarityhub/frontend/model/enums/NeedType.java
@@ -1,0 +1,5 @@
+package solidarityhub.frontend.model.enums;
+
+public enum NeedType {
+    MEDICAL, POLICE, FIREFIGHTERS, CLEANING, FEED, PSYCHOLOGICAL, BUILDING, CLOTHING, REFUGE, OTHER
+}
