+++ Frontend/src/main/java/solidarityhub/frontend/model/enums/Priority.java
@@ -1,0 +1,5 @@
+package solidarityhub.frontend.model.enums;
+
+public enum Priority {
+    URGENT, MODERATE, LOW
+}
