+++ Frontend/src/main/java/solidarityhub/frontend/model/enums/Status.java
@@ -1,0 +1,5 @@
+package solidarityhub.frontend.model.enums;
+
+public enum Status {
+    TO_DO, IN_PROGRESS, FINISHED
+}
