+++ Frontend/src/main/java/solidarityhub/frontend/model/enums/UrgencyLevel.java
@@ -1,0 +1,5 @@
+package solidarityhub.frontend.model.enums;
+
+public enum UrgencyLevel {
+    URGENT, MODERATE, LOW
+}
